Added anti affinity features.
Abiquo 2.4 introduces the concept of Layers in a virtual
appliance. This allows virtual machines deployed in the same layer
to be configured to use the anti affinity features and force
them to be deployed in different hosts.

This is useful for users deploying high available services, to
make sure that a failure in a host will not make the service crash.