Added template definition entity.
In Abiquo, template definitions are OVF representations of the templates
that can be made available to each tenant. Those definitions contain
information such as the operating system type, architecture, default
login credentials, etc.

This commit adds the Api calls needed to manage the templates to make
them available to the tenants, and also improvements to the Abiquo
ComputeService implementation that makes use of all the information
provided by the template definitions.