Enable volume and hard disk ordering
Since Abiquo 2.4 HardDisk and Volume attachment order can be configured
in virtual machines. This way, the user will be able to know in advance
in which device of the virtual machine the virtual disk will be
attached and will be able to properly build scripts that access the
external virtual disks.