Added generic refresh method in Abiquo
In Abiquo every object contains a link to itself with rel="edit". This
commit adds a generic method in the DomainWrapper base class that reads
the link and performs a GET operation to refresh the object.

This method was already used in conversions and virtual machines, and
now every object will have it implemented.